Fixed type on rating table

+++ tables/rating.sql
@@ -1,11 +1,11 @@
 DROP TABLE IF EXISTS rating;
 CREATE TABLE rating (
     post_id INT NOT NULL,
-    attitude  DECIMAL(2,1) NOT NULL CHECK (attitude >= 0 AND attitude <= 5),
-    communication DECIMAL(2,1) NOT NULL CHECK (communication >= 0 AND communication <= 5),
-    efficiency DECIMAL(2,1) NOT NULL CHECK (efficiency >= 0 AND efficiency <= 5),
-    fairness DECIMAL(2,1) NOT NULL CHECK (fairness >= 0 AND fairness <= 5),
-    safety DECIMAL(2,1) NOT NULL CHECK (safety >= 0 AND safety <= 5),
+    attitude  INT NOT NULL CHECK (attitude >= 0 AND attitude <= 5),
+    communication INT NOT NULL CHECK (communication >= 0 AND communication <= 5),
+    efficiency INT NOT NULL CHECK (efficiency >= 0 AND efficiency <= 5),
+    fairness INT NOT NULL CHECK (fairness >= 0 AND fairness <= 5),
+    safety INT NOT NULL CHECK (safety >= 0 AND safety <= 5),
     PRIMARY KEY (post_id)
 );
